split useRecentWorldUsers into it's own file
