+++ src/hooks/users/index.ts
@@ -1,8 +1,5 @@
 import { useMemo } from "react";
 
-import { User } from "types/User";
-
-import { WithId } from "utils/id";
 import { normalizeTimestampToMilliseconds } from "utils/time";
 
 import { useConnectCurrentVenueNG } from "hooks/useConnectCurrentVenueNG";
@@ -17,26 +14,7 @@ export {
   useWorldUsersById,
   useWorldUsersByIdWorkaround,
 } from "./useWorldUsersById";
-
-export const useRecentWorldUsers = (): {
-  recentWorldUsers: readonly WithId<User>[];
-  isRecentWorldUsersLoaded: boolean;
-} => {
-  const lastSeenThreshold = useUserLastSeenThreshold();
-
-  const { worldUsers, isWorldUsersLoaded } = useWorldUsers();
-
-  return useMemo(
-    () => ({
-      recentWorldUsers: worldUsers.filter(
-        (user) =>
-          normalizeTimestampToMilliseconds(user.lastSeenAt) > lastSeenThreshold
-      ),
-      isRecentWorldUsersLoaded: isWorldUsersLoaded,
-    }),
-    [worldUsers, isWorldUsersLoaded, lastSeenThreshold]
-  );
-};
+export { useRecentWorldUsers } from "./useRecentWorldUsers";
 
 /**
  * @description this hook's filtering world users based on their @lastSeenIn location
