+++ src/hooks/users/useRecentWorldUsers.ts
@@ -1,0 +1,30 @@
+import { useMemo } from "react";
+
+import { User } from "types/User";
+
+import { WithId } from "utils/id";
+import { normalizeTimestampToMilliseconds } from "utils/time";
+
+import { useUserLastSeenThreshold } from "hooks/useUserLastSeenThreshold";
+
+import { useWorldUsers } from "./useWorldUsers";
+
+export const useRecentWorldUsers = (): {
+  recentWorldUsers: readonly WithId<User>[];
+  isRecentWorldUsersLoaded: boolean;
+} => {
+  const lastSeenThreshold = useUserLastSeenThreshold();
+
+  const { worldUsers, isWorldUsersLoaded } = useWorldUsers();
+
+  return useMemo(
+    () => ({
+      recentWorldUsers: worldUsers.filter(
+        (user) =>
+          normalizeTimestampToMilliseconds(user.lastSeenAt) > lastSeenThreshold
+      ),
+      isRecentWorldUsersLoaded: isWorldUsersLoaded,
+    }),
+    [worldUsers, isWorldUsersLoaded, lastSeenThreshold]
+  );
+};
